Fix wedge_power.sh do_on's logic to call improve_aura_pll
Summary:
[FUJI] Fix wedge_power.sh do_on's logic to call
improve_aura_pll
Currently, when wedge_power.sh on is executed with -f,
improve_aura_pll.sh is executed no matter what, even when CPU is already
on.
Problem is, when BMC boots up, it tries to turn on CPU with -f option.
This results in the execution of improve_aura_pll.sh regardless of the
CPU's power-on state. This diff fixes this issue by running
improve_aura_pll.sh only when CPU is off, even when -f option is
applied.

Test Plan:
BEFORE : P1674339966
AFTER : P1674339003

Reviewed By: alandau

Differential Revision: D65554662

fbshipit-source-id: 641f12716e94ffa219ddabe1335570220ba78453