yosemite4: fw-versions …
Summary:
# Description:
There is a timeout check when execute the cxl-fw-update. Since the script is set -e, the script will be terminated if timeout occurs. Revise the script to check the version string instead of checking timeout command.

# Motivation:
The script is not supposed to be terminated if the cxl-fw-update command timeout.

X-link: https://github.com/facebookexternal/openbmc.wiwynn/pull/3234

Test Plan:
Build code - pass
Tested on YV4 - pass

Reviewed By: Prabha-Veerubhotla

Differential Revision: D67013518

fbshipit-source-id: 4092644b02a65cdb24e75a5018016ab58962a959